+++ src/crimson/os/seastore/CMakeLists.txt
@@ -12,8 +12,6 @@ set(crimson_seastore_srcs
   lba_manager.cc
   segment_cleaner.cc
   lba_manager/btree/btree_lba_manager.cc
-  lba_manager/btree/btree_range_pin.cc
-  lba_manager/btree/lba_btree.cc
   lba_manager/btree/lba_btree_node.cc
   omap_manager.cc
   omap_manager/btree/btree_omap_manager.cc
