+++ src/crimson/os/seastore/lba_manager/btree/btree_lba_manager.h
@@ -15,15 +15,18 @@
 #include "common/interval_map.h"
 #include "crimson/osd/exceptions.h"
 
+#include "crimson/os/seastore/btree/fixed_kv_btree.h"
 #include "crimson/os/seastore/seastore_types.h"
 #include "crimson/os/seastore/lba_manager.h"
 #include "crimson/os/seastore/cache.h"
 #include "crimson/os/seastore/segment_manager.h"
 
 #include "crimson/os/seastore/lba_manager/btree/lba_btree_node.h"
-#include "crimson/os/seastore/lba_manager/btree/lba_btree.h"
+#include "crimson/os/seastore/btree/btree_range_pin.h"
 
 namespace crimson::os::seastore::lba_manager::btree {
+
+using LBABtree = FixedKVBtree<laddr_t, lba_map_val_t, LBAInternalNode, LBALeafNode, LBA_BLOCK_SIZE>;
 
 /**
  * BtreeLBAManager
@@ -84,6 +87,14 @@ public:
   void complete_transaction(
     Transaction &t) final;
 
+  /**
+   * init_cached_extent
+   *
+   * Checks whether e is live (reachable from lba tree) and drops or initializes
+   * accordingly.
+   *
+   * Returns if e is live.
+   */
   init_cached_extent_ret init_cached_extent(
     Transaction &t,
     CachedExtentRef e) final;
@@ -117,8 +128,8 @@ public:
 
   void add_pin(LBAPin &pin) final {
     auto *bpin = reinterpret_cast<BtreeLBAPin*>(&pin);
-    pin_set.add_pin(bpin->pin);
-    bpin->parent = nullptr;
+    pin_set.add_pin(bpin->get_range_pin());
+    bpin->set_parent(nullptr);
   }
 
   ~BtreeLBAManager();
@@ -126,24 +137,24 @@ private:
   SegmentManager &segment_manager;
   Cache &cache;
 
-  btree_pin_set_t pin_set;
+  btree_pin_set_t<laddr_t> pin_set;
 
   struct {
     uint64_t num_alloc_extents = 0;
     uint64_t num_alloc_extents_iter_nexts = 0;
   } stats;
 
-  op_context_t get_context(Transaction &t) {
-    return op_context_t{cache, t, &pin_set};
+  op_context_t<laddr_t> get_context(Transaction &t) {
+    return op_context_t<laddr_t>{cache, t, &pin_set};
   }
 
-  static btree_range_pin_t &get_pin(CachedExtent &e);
+  static btree_range_pin_t<laddr_t> &get_pin(CachedExtent &e);
 
   seastar::metrics::metric_group metrics;
   void register_metrics();
   template <typename F, typename... Args>
   auto with_btree(
-    op_context_t c,
+    op_context_t<laddr_t> c,
     F &&f) {
     return cache.get_root(
       c.trans
@@ -168,7 +179,7 @@ private:
 
   template <typename State, typename F>
   auto with_btree_state(
-    op_context_t c,
+    op_context_t<laddr_t> c,
     State &&init,
     F &&f) {
     return seastar::do_with(
@@ -185,14 +196,14 @@ private:
 
   template <typename State, typename F>
   auto with_btree_state(
-    op_context_t c,
+    op_context_t<laddr_t> c,
     F &&f) {
     return with_btree_state<State, F>(c, State{}, std::forward<F>(f));
   }
 
   template <typename Ret, typename F>
   auto with_btree_ret(
-    op_context_t c,
+    op_context_t<laddr_t> c,
     F &&f) {
     return with_btree_state<Ret>(
       c,
